# diplom/cashFlowManager/cfm-api/swagger-api-specification/api-specification.yaml
openapi: 3.0.0
info:
  title: Cash Flow Manager API
  version: 0.0.1
...
paths:
...
  /account/list/{ownerId}/:
    get:
      tags:
        - account
      summary: Получение всех счетов пользователя
      operationId: getAccountByOwnerList
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountByOwnerListResponse'
        '400':
          description: Некорректный запрос от клиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизованный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Запрашиваемый ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /account/list/:
    get:
      tags:
        - account
      summary: Получение списка всех счетов
      operationId: getAccountsList
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsListResponse'
        '400':
          description: Некорректный запрос от клиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизованный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Запрашиваемый ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /account/add/:
    post:
      tags:
        - account
      summary: Создание нового счета
      operationId: addAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAccountRequest'
      responses:
        '200':
          description: Успешное создание счета
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddAccountResponse'
        '400':
          description: Некорректный запрос от клиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизованный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Запрашиваемый ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /account/{accountId}/:
    get:
      tags:
        - account
      summary: Получение информации о конкретном счете
      operationId: getAccountInfo
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountInfoResponse'
        '400':
          description: Некорректный запрос от клиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизованный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Запрашиваемый ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /account/update/{accountId}/:
    post:
      tags:
        - account
      summary: Обновление информации о счете
      operationId: updateAccount
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountRequest'
      responses:
        '200':
          description: Успешное обновление счета
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateAccountResponse'
        '400':
          description: Некорректный запрос от клиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизованный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Запрашиваемый ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /account/delete/{accountId}/:
    post:
      tags:
        - account
      summary: Удаление счета
      operationId: deleteAccount
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успешное удаление счета
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAccountResponse'
        '400':
          description: Некорректный запрос от клиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизованный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Запрашиваемый ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /account/deleted/list/:
    get:
      tags:
        - account
      summary: Получение списка всех удаленных счетов
      operationId: getDeletedAccountsList
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedAccountsListResponse'
        '400':
          description: Некорректный запрос от клиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизованный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Запрашиваемый ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      description: Ошибка от приложения
      type: object
      required:
        - status
        - error
        - message
        - exception
        - path
        - timestamp
      properties:
        status:
          description: Статус ошибки
          type: string
          example: "500"
          pattern: "^[45]0[0-9]{1}$"  # Ошибка сервера (5xx) или клиента (4xx)
          maxLength: 3  # Максимальная длина строки
        error:
          description: Описание ошибки
          type: string
          example: "Internal Server Error"
        message:
          description: Сообщение об ошибке
          type: string
          example: "Произошла ошибка при обработке запроса"
        exception:
          description: Имя исключения
          type: string
          example: "NullPointerException"
        path:
          description: Путь запроса, вызвавшего ошибку
          type: string
          example: "/api/v1/users/123"
        timestamp:
          description: Временная метка ошибки
          type: string
          format: date-time  # Формат даты-времени
          example: "2024-02-06T12:30:45Z"
      additionalProperties: false
...
    AccountItem:
      description: "Элемент данных о счете"
      type: "object"
      required:
        - id
        - accountNumber
        - accountOwnerId
        - balance
        - currency
        - status
        - createdAt
      properties:
        id:
          description: "Идентификатор счета"
          type: "string"
        accountNumber:
          description: "Номер счета"
          type: "string"
        accountOwnerId:
          description: "Идентификатор владельца счета"
          type: "string"
        balance:
          description: "Баланс счета"
          type: "integer"
        currency:
          description: "Валюта счета"
          type: "string"
        status:
          description: "Статус счета"
          type: "string"
        createdAt:
          description: "Дата создания счета"
          type: integer
          format: int64
      additionalProperties: false
...
    AccountsListResponse:
      description: "Ответ со списком всех счетов"
      type: "object"
      required:
        - accountsList
      properties:
        'accountsList':
          description: "Список всех счетов"
          type: array
          items:
            $ref: '#/components/schemas/AccountItem'
      additionalProperties: false
...
    AccountByOwnerListResponse:
      description: "Ответ со списком всех счетов пользователя"
      type: "object"
      required:
        - accountsList
      properties:
        'accountsList':
          description: "Список всех счетов пользователя"
          type: array
          items:
            $ref: '#/components/schemas/AccountItem'
      additionalProperties: false
...
    AddAccountRequest:
      description: "Запрос на добавление счета"
      required:
        - account
      type: "object"
      properties:
        'account':
          $ref: '#/components/schemas/AccountItem'
    AddAccountResponse:
      description: "Ответ на добавление счета"
      type: "object"
      required:
        - accountId
      properties:
        accountId:
          description: Идентификатор только что созданного счета
          type: "string"
    AccountInfoResponse:
      description: "Ответ на получение информации о счете"
      type: "object"
      required:
        - account
      properties:
        'account':
          $ref: '#/components/schemas/AccountItem'
    UpdateAccountRequest:
      description: "Запрос на изменение информации о счете"
      type: "object"
      required:
        - account
      properties:
        'account':
          $ref: '#/components/schemas/AccountItem'
    UpdateAccountResponse:
      description: "Ответ на изменение информации о счете"
      type: "object"
      required:
        - account
      properties:
        'account':
          $ref: '#/components/schemas/AccountItem'
    DeleteAccountResponse:
      description: "Ответ на удаление счета"
      type: "object"
      required:
        - accountId
      properties:
        accountId:
          description: Идентификатор только что удаленного счета
          type: "string"
...
    DeletedAccountsListResponse:
      description: "Ответ со списком всех удаленных счетов"
      type: "object"
      required:
        - deletedAccountsList
      properties:
        'deletedAccountsList':
          description: "Список всех удаленных счетов"
          type: array
          items:
            $ref: '#/components/schemas/AccountItem'